+++ persist-cli/src/main/java/io/ballerina/persist/nodegenerator/syntax/clients/GSheetClientSyntax.java
@@ -39,6 +39,9 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 
+import static io.ballerina.persist.nodegenerator.syntax.constants.BalSyntaxConstants.GOOGLE_SHEETS;
+import static io.ballerina.persist.nodegenerator.syntax.utils.BalSyntaxUtils.EntityType.SHEET;
+
 /**
  * This class is used to generate the DB client syntax tree.
  *
@@ -46,6 +49,7 @@ import java.util.Map;
  */
 public class GSheetClientSyntax implements ClientSyntax {
 
+    public static final String GOOGLE_SHEETS_PROCESSOR = "GoogleSheetsProcessor";
     private final Map<String, String> queryMethodStatement = new HashMap<>();
     private final Module entityModule;
     public GSheetClientSyntax(Module entityModule) {
@@ -59,7 +63,7 @@ public class GSheetClientSyntax implements ClientSyntax {
                 BalSyntaxConstants.GOOGLE_API_SHEET, null));
         imports = imports.add(BalSyntaxUtils.getImportDeclarationNode(BalSyntaxConstants.KEYWORD_BALLERINAX,
                 BalSyntaxConstants.PERSIST_MODULE + "." +
-                        BalSyntaxConstants.GOOGLE_SHEETS, null));
+                        GOOGLE_SHEETS, null));
         return imports;
     }
 
@@ -70,7 +74,7 @@ public class GSheetClientSyntax implements ClientSyntax {
 
     @Override
     public Client getClientObject(Module entityModule, String clientName) {
-        Client clientObject = BalSyntaxUtils.generateClientSignature(clientName, true);
+        Client clientObject = BalSyntaxUtils.generateClientSignature(clientName, true, GOOGLE_SHEETS);
         clientObject.addMember(NodeParser.parseObjectMember(BalSyntaxConstants.GOOGLE_SHEET_CLIENT), true);
         clientObject.addMember(NodeParser.parseObjectMember(BalSyntaxConstants.HTTP_CLIENT), true);
         clientObject.addMember(NodeParser.parseObjectMember(BalSyntaxConstants.INIT_GOOGLE_SHEET_CLIENT_MAP), true);
@@ -118,19 +122,27 @@ public class GSheetClientSyntax implements ClientSyntax {
 
     @Override
     public FunctionDefinitionNode getGetFunction(Entity entity) {
-        return BalSyntaxUtils.generateGetFunction(entity, "GoogleSheetsProcessor",
-                BalSyntaxConstants.GOOGLE_SHEETS);
+        FunctionDefinitionNode functionNode = BalSyntaxUtils.generateGetFunction(entity, GOOGLE_SHEETS_PROCESSOR,
+                GOOGLE_SHEETS);
+
+        String doc = BalSyntaxUtils.createGetResourceDocumentation(entity, SHEET, false);
+        return BalSyntaxUtils.addDocumentationToFunction(functionNode, doc);
     }
 
     @Override
     public FunctionDefinitionNode getGetByKeyFunction(Entity entity) {
-        return BalSyntaxUtils.generateGetByKeyFunction(entity, "GoogleSheetsProcessor",
-                BalSyntaxConstants.GOOGLE_SHEETS);
+        FunctionDefinitionNode functionNode = BalSyntaxUtils.generateGetByKeyFunction(entity, GOOGLE_SHEETS_PROCESSOR,
+                GOOGLE_SHEETS);
+
+        String doc = BalSyntaxUtils.createGetByKeyResourceDocumentation(entity, SHEET);
+        return BalSyntaxUtils.addDocumentationToFunction(functionNode, doc);
     }
 
     @Override
     public FunctionDefinitionNode getCloseFunction() {
         Function close = BalSyntaxUtils.generateCloseFunction();
+        String doc = BalSyntaxUtils.createCloseMethodDocumentation();
+        close.addDocumentation(BalSyntaxUtils.createMarkdownDocumentationNode(doc));
         close.addStatement(NodeParser.parseStatement(BalSyntaxConstants.RETURN_NIL));
         return close.getFunctionDefinitionNode();
     }
@@ -140,6 +152,8 @@ public class GSheetClientSyntax implements ClientSyntax {
         String parameterType = String.format(BalSyntaxConstants.INSERT_RECORD, entity.getEntityName());
         List<EntityField> primaryKeys = entity.getKeys();
         Function create = BalSyntaxUtils.generatePostFunction(entity, primaryKeys, parameterType);
+        String doc = BalSyntaxUtils.createPostResourceDocumentation(entity, SHEET);
+        create.addDocumentation(BalSyntaxUtils.createMarkdownDocumentationNode(doc));
         addFunctionBodyToPostResource(create, primaryKeys,
                 BalSyntaxUtils.getStringWithUnderScore(entity.getEntityName()), parameterType);
         return create.getFunctionDefinitionNode();
@@ -150,6 +164,8 @@ public class GSheetClientSyntax implements ClientSyntax {
         StringBuilder filterKeys = new StringBuilder(BalSyntaxConstants.OPEN_BRACE);
         StringBuilder path = new StringBuilder(BalSyntaxConstants.BACK_SLASH + entity.getClientResourceName());
         Function update = BalSyntaxUtils.generatePutFunction(entity, filterKeys, path);
+        String doc = BalSyntaxUtils.createPutResourceDocumentation(entity, SHEET);
+        update.addDocumentation(BalSyntaxUtils.createMarkdownDocumentationNode(doc));
         if (entity.getKeys().size() > 1) {
 
             update.addStatement(NodeParser.parseStatement(BalSyntaxConstants.G_SHEET_CLIENT_DECLARATION));
@@ -182,6 +198,8 @@ public class GSheetClientSyntax implements ClientSyntax {
         StringBuilder filterKeys = new StringBuilder(BalSyntaxConstants.OPEN_BRACE);
         StringBuilder path = new StringBuilder(BalSyntaxConstants.BACK_SLASH + entity.getClientResourceName());
         Function delete = BalSyntaxUtils.generateDeleteFunction(entity, filterKeys, path);
+        String doc = BalSyntaxUtils.createDeleteResourceDocumentation(entity, SHEET);
+        delete.addDocumentation(BalSyntaxUtils.createMarkdownDocumentationNode(doc));
         delete.addStatement(NodeParser.parseStatement(String.format(BalSyntaxConstants.GET_OBJECT_QUERY,
                 entity.getEntityName(), path)));
         if (entity.getKeys().size() > 1) {